refactor xds code
