+++ pilot/pkg/xds/ads.go
@@ -68,15 +68,15 @@ type DeltaDiscoveryClient = discovery.AggregatedDiscoveryService_DeltaAggregated
 
 // Connection holds information about connected client.
 type Connection struct {
-	// PeerAddr is the address of the client, from network layer.
-	PeerAddr string
+	// peerAddr is the address of the client, from network layer.
+	peerAddr string
 
 	// Time of connection, for debugging
-	Connect time.Time
+	connectedAt time.Time
 
-	// ConID is the connection identifier, used as a key in the connection table.
+	// identifier is the connection identifier, used as a key in the connection table.
 	// Currently based on the node name and a counter.
-	ConID string
+	identifier string
 
 	// proxy is the client to which this connection is established.
 	proxy *model.Proxy
@@ -124,8 +124,8 @@ func newConnection(peerAddr string, stream DiscoveryStream) *Connection {
 		stop:        make(chan struct{}),
 		reqChan:     make(chan *discovery.DiscoveryRequest, 1),
 		errorChan:   make(chan error, 1),
-		PeerAddr:    peerAddr,
-		Connect:     time.Now(),
+		peerAddr:    peerAddr,
+		connectedAt: time.Now(),
 		stream:      stream,
 	}
 }
@@ -147,11 +147,11 @@ func (s *DiscoveryServer) receive(con *Connection, identities []string) {
 		req, err := con.stream.Recv()
 		if err != nil {
 			if istiogrpc.IsExpectedGRPCError(err) {
-				log.Infof("ADS: %q %s terminated %v", con.PeerAddr, con.ConID, err)
+				log.Infof("ADS: %q %s terminated", con.peerAddr, con.identifier)
 				return
 			}
 			con.errorChan <- err
-			log.Errorf("ADS: %q %s terminated with error: %v", con.PeerAddr, con.ConID, err)
+			log.Errorf("ADS: %q %s terminated with error: %v", con.peerAddr, con.identifier, err)
 			totalXDSInternalErrors.Increment()
 			return
 		}
@@ -159,7 +159,7 @@ func (s *DiscoveryServer) receive(con *Connection, identities []string) {
 		if firstRequest {
 			// probe happens before envoy sends first xDS request
 			if req.TypeUrl == v3.HealthInfoType {
-				log.Warnf("ADS: %q %s send health check probe before normal xDS request", con.PeerAddr, con.ConID)
+				log.Warnf("ADS: %q %s send health check probe before normal xDS request", con.peerAddr, con.identifier)
 				continue
 			}
 			firstRequest = false
@@ -172,13 +172,13 @@ func (s *DiscoveryServer) receive(con *Connection, identities []string) {
 				return
 			}
 			defer s.closeConnection(con)
-			log.Infof("ADS: new connection for node:%s", con.ConID)
+			log.Infof("ADS: new connection for node:%s", con.identifier)
 		}
 
 		select {
 		case con.reqChan <- req:
 		case <-con.stream.Context().Done():
-			log.Infof("ADS: %q %s terminated with stream closed", con.PeerAddr, con.ConID)
+			log.Infof("ADS: %q %s terminated with stream closed", con.peerAddr, con.identifier)
 			return
 		}
 	}
@@ -188,7 +188,8 @@ func (s *DiscoveryServer) receive(con *Connection, identities []string) {
 // handles 'push' requests and close - the code will eventually call the 'push' code, and it needs more mutex
 // protection. Original code avoided the mutexes by doing both 'push' and 'process requests' in same thread.
 func (s *DiscoveryServer) processRequest(req *discovery.DiscoveryRequest, con *Connection) error {
-	if !s.shouldProcessRequest(con.proxy, req) {
+	if req.TypeUrl == v3.HealthInfoType {
+		s.handleHealthRequest(con.proxy, req)
 		return nil
 	}
 
@@ -199,7 +200,7 @@ func (s *DiscoveryServer) processRequest(req *discovery.DiscoveryRequest, con *C
 			&model.PushRequest{Full: true, Push: con.proxy.LastPushContext})
 	}
 	if s.StatusReporter != nil {
-		s.StatusReporter.RegisterEvent(con.ConID, req.TypeUrl, req.ResponseNonce)
+		s.StatusReporter.RegisterEvent(con.identifier, req.TypeUrl, req.ResponseNonce)
 	}
 	shouldRespond := s.shouldRespond(con, req)
 	if !shouldRespond {
@@ -258,11 +259,6 @@ func (s *DiscoveryServer) Stream(stream DiscoveryStream) error {
 		peerAddr = peerInfo.Addr.String()
 	}
 
-	if err := s.WaitForRequestLimit(stream.Context()); err != nil {
-		log.Warnf("ADS: %q exceeded rate limit: %v", peerAddr, err)
-		return status.Errorf(codes.ResourceExhausted, "request rate limit exceeded: %v", err)
-	}
-
 	ids, err := s.authenticate(ctx)
 	if err != nil {
 		return status.Error(codes.Unauthenticated, err.Error())
@@ -271,6 +267,11 @@ func (s *DiscoveryServer) Stream(stream DiscoveryStream) error {
 		log.Debugf("Authenticated XDS: %v with identity %v", peerAddr, ids)
 	} else {
 		log.Debugf("Unauthenticated XDS: %s", peerAddr)
+	}
+
+	if err := s.WaitForRequestLimit(stream.Context()); err != nil {
+		log.Warnf("ADS: %q exceeded rate limit: %v", peerAddr, err)
+		return status.Errorf(codes.ResourceExhausted, "request rate limit exceeded: %v", err)
 	}
 
 	// InitContext returns immediately if the context was already initialized.
@@ -331,7 +332,7 @@ func (s *DiscoveryServer) shouldRespond(con *Connection, request *discovery.Disc
 	// will be different from the version sent. But it is fragile to rely on that.
 	if request.ErrorDetail != nil {
 		errCode := codes.Code(request.ErrorDetail.Code)
-		log.Warnf("ADS:%s: ACK ERROR %s %s:%s", stype, con.ConID, errCode.String(), request.ErrorDetail.GetMessage())
+		log.Warnf("ADS:%s: ACK ERROR %s %s:%s", stype, con.identifier, errCode.String(), request.ErrorDetail.GetMessage())
 		incrementXDSRejects(request.TypeUrl, con.proxy.ID, errCode.String())
 		if s.StatusGen != nil {
 			s.StatusGen.OnNack(con.proxy, request)
@@ -345,7 +346,7 @@ func (s *DiscoveryServer) shouldRespond(con *Connection, request *discovery.Disc
 	}
 
 	if shouldUnsubscribe(request) {
-		log.Debugf("ADS:%s: UNSUBSCRIBE %s %s %s", stype, con.ConID, request.VersionInfo, request.ResponseNonce)
+		log.Debugf("ADS:%s: UNSUBSCRIBE %s %s %s", stype, con.identifier, request.VersionInfo, request.ResponseNonce)
 		con.proxy.Lock()
 		delete(con.proxy.WatchedResources, request.TypeUrl)
 		con.proxy.Unlock()
@@ -363,7 +364,7 @@ func (s *DiscoveryServer) shouldRespond(con *Connection, request *discovery.Disc
 	// because Istiod is restarted or Envoy disconnects and reconnects.
 	// We should always respond with the current resource names.
 	if request.ResponseNonce == "" || previousInfo == nil {
-		log.Debugf("ADS:%s: INIT/RECONNECT %s %s %s", stype, con.ConID, request.VersionInfo, request.ResponseNonce)
+		log.Debugf("ADS:%s: INIT/RECONNECT %s %s %s", stype, con.identifier, request.VersionInfo, request.ResponseNonce)
 		con.proxy.Lock()
 		con.proxy.WatchedResources[request.TypeUrl] = &model.WatchedResource{TypeUrl: request.TypeUrl, ResourceNames: request.ResourceNames}
 		con.proxy.Unlock()
@@ -374,7 +375,7 @@ func (s *DiscoveryServer) shouldRespond(con *Connection, request *discovery.Disc
 	// A nonce becomes stale following a newer nonce being sent to Envoy.
 	if request.ResponseNonce != previousInfo.NonceSent {
 		log.Debugf("ADS:%s: REQ %s Expired nonce received %s, sent %s", stype,
-			con.ConID, request.ResponseNonce, previousInfo.NonceSent)
+			con.identifier, request.ResponseNonce, previousInfo.NonceSent)
 		xdsExpiredNonce.With(typeTag.Value(v3.GetMetricType(request.TypeUrl))).Increment()
 		con.proxy.Lock()
 		con.proxy.WatchedResources[request.TypeUrl].NonceNacked = ""
@@ -394,11 +395,11 @@ func (s *DiscoveryServer) shouldRespond(con *Connection, request *discovery.Disc
 	// Envoy can send two DiscoveryRequests with same version and nonce
 	// when it detects a new resource. We should respond if they change.
 	if listEqualUnordered(previousResources, request.ResourceNames) {
-		log.Debugf("ADS:%s: ACK %s %s %s", stype, con.ConID, request.VersionInfo, request.ResponseNonce)
+		log.Debugf("ADS:%s: ACK %s %s %s", stype, con.identifier, request.VersionInfo, request.ResponseNonce)
 		return false
 	}
 	log.Debugf("ADS:%s: RESOURCE CHANGE previous resources: %v, new resources: %v %s %s %s", stype,
-		previousResources, request.ResourceNames, con.ConID, request.VersionInfo, request.ResponseNonce)
+		previousResources, request.ResourceNames, con.identifier, request.VersionInfo, request.ResponseNonce)
 
 	return true
 }
@@ -467,7 +468,7 @@ func (s *DiscoveryServer) initConnection(node *core.Node, con *Connection, ident
 	// way only new push contexts will be registered for this proxy.
 	proxy.LastPushContext = s.globalPushContext()
 	// First request so initialize connection id and start tracking it.
-	con.ConID = connectionID(proxy.ID)
+	con.identifier = connectionID(proxy.ID)
 	con.node = node
 	con.proxy = proxy
 
@@ -481,7 +482,7 @@ func (s *DiscoveryServer) initConnection(node *core.Node, con *Connection, ident
 	// a better choice, it introduces a race condition; If we complete initialization of a new push
 	// context between initializeProxy and addCon, we would not get any pushes triggered for the new
 	// push context, leading the proxy to have a stale state until the next full push.
-	s.addCon(con.ConID, con)
+	s.addCon(con.identifier, con)
 	// Register that initialization is complete. This triggers to calls that it is safe to access the
 	// proxy
 	defer close(con.initialized)
@@ -499,17 +500,17 @@ func (s *DiscoveryServer) initConnection(node *core.Node, con *Connection, ident
 }
 
 func (s *DiscoveryServer) closeConnection(con *Connection) {
-	if con.ConID == "" {
+	if con.identifier == "" {
 		return
 	}
-	s.removeCon(con.ConID)
+	s.removeCon(con.identifier)
 	if s.StatusGen != nil {
 		s.StatusGen.OnDisconnect(con)
 	}
 	if s.StatusReporter != nil {
-		s.StatusReporter.RegisterDisconnect(con.ConID, AllEventTypesList)
+		s.StatusReporter.RegisterDisconnect(con.identifier, AllEventTypesList)
 	}
-	s.WorkloadEntryController.QueueUnregisterWorkload(con.proxy, con.Connect)
+	s.WorkloadEntryController.QueueUnregisterWorkload(con.proxy, con.connectedAt)
 }
 
 func connectionID(node string) string {
@@ -541,7 +542,7 @@ func (s *DiscoveryServer) initializeProxy(node *core.Node, con *Connection) erro
 	proxy := con.proxy
 	// this should be done before we look for service instances, but after we load metadata
 	// TODO fix check in kubecontroller treat echo VMs like there isn't a pod
-	if err := s.WorkloadEntryController.RegisterWorkload(proxy, con.Connect); err != nil {
+	if err := s.WorkloadEntryController.RegisterWorkload(proxy, con.connectedAt); err != nil {
 		return err
 	}
 	s.computeProxyState(proxy, nil)
@@ -641,11 +642,8 @@ func (s *DiscoveryServer) computeProxyState(proxy *model.Proxy, request *model.P
 	}
 }
 
-// shouldProcessRequest returns whether or not to continue with the request.
-func (s *DiscoveryServer) shouldProcessRequest(proxy *model.Proxy, req *discovery.DiscoveryRequest) bool {
-	if req.TypeUrl != v3.HealthInfoType {
-		return true
-	}
+// handleHealthRequest processes HealthInformation type Url.
+func (s *DiscoveryServer) handleHealthRequest(proxy *model.Proxy, req *discovery.DiscoveryRequest) {
 	if features.WorkloadEntryHealthChecks {
 		event := workloadentry.HealthEvent{}
 		event.Healthy = req.ErrorDetail == nil
@@ -654,7 +652,6 @@ func (s *DiscoveryServer) shouldProcessRequest(proxy *model.Proxy, req *discover
 		}
 		s.WorkloadEntryController.QueueWorkloadEntryHealth(proxy, event)
 	}
-	return false
 }
 
 // DeltaAggregatedResources is not implemented.
@@ -678,10 +675,10 @@ func (s *DiscoveryServer) pushConnection(con *Connection, pushEv *Event) error {
 	}
 
 	if !s.ProxyNeedsPush(con.proxy, pushRequest) {
-		log.Debugf("Skipping push to %v, no updates required", con.ConID)
+		log.Debugf("Skipping push to %v, no updates required", con.identifier)
 		if pushRequest.Full {
 			// Only report for full versions, incremental pushes do not have a new version.
-			reportAllEvents(s.StatusReporter, con.ConID, pushRequest.Push.LedgerVersion, nil)
+			reportAllEvents(s.StatusReporter, con.identifier, pushRequest.Push.LedgerVersion, nil)
 		}
 		return nil
 	}
@@ -696,7 +693,7 @@ func (s *DiscoveryServer) pushConnection(con *Connection, pushEv *Event) error {
 	}
 	if pushRequest.Full {
 		// Report all events for unwatched resources. Watched resources will be reported in pushXds or on ack.
-		reportAllEvents(s.StatusReporter, con.ConID, pushRequest.Push.LedgerVersion, ignoreEvents)
+		reportAllEvents(s.StatusReporter, con.identifier, pushRequest.Push.LedgerVersion, ignoreEvents)
 	}
 
 	proxiesConvergeDelay.Record(time.Since(pushRequest.Start).Seconds())
@@ -856,7 +853,7 @@ func (conn *Connection) send(res *discovery.DiscoveryResponse) error {
 			conn.proxy.Unlock()
 		}
 	} else if status.Convert(err).Code() == codes.DeadlineExceeded {
-		log.Infof("Timeout writing %s", conn.ConID)
+		log.Infof("Timeout writing %s", conn.identifier)
 		xdsResponseWriteTimeouts.Increment()
 	}
 	return err
