+++ pilot/pkg/xds/delta.go
@@ -28,7 +28,6 @@ import (
 	"istio.io/istio/pilot/pkg/features"
 	istiogrpc "istio.io/istio/pilot/pkg/grpc"
 	"istio.io/istio/pilot/pkg/model"
-	"istio.io/istio/pilot/pkg/networking/util"
 	"istio.io/istio/pilot/pkg/util/sets"
 	v3 "istio.io/istio/pilot/pkg/xds/v3"
 	istiolog "istio.io/pkg/log"
@@ -138,10 +137,10 @@ func (s *DiscoveryServer) pushConnectionDelta(con *Connection, pushEv *Event) er
 	}
 
 	if !s.ProxyNeedsPush(con.proxy, pushRequest) {
-		deltaLog.Debugf("Skipping push to %v, no updates required", con.ConID)
+		deltaLog.Debugf("Skipping push to %v, no updates required", con.identifier)
 		if pushRequest.Full {
 			// Only report for full versions, incremental pushes do not have a new version
-			reportAllEvents(s.StatusReporter, con.ConID, pushRequest.Push.LedgerVersion, nil)
+			reportAllEvents(s.StatusReporter, con.identifier, pushRequest.Push.LedgerVersion, nil)
 		}
 		return nil
 	}
@@ -156,7 +155,7 @@ func (s *DiscoveryServer) pushConnectionDelta(con *Connection, pushEv *Event) er
 	}
 	if pushRequest.Full {
 		// Report all events for unwatched resources. Watched resources will be reported in pushXds or on ack.
-		reportAllEvents(s.StatusReporter, con.ConID, pushRequest.Push.LedgerVersion, ignoreEvents)
+		reportAllEvents(s.StatusReporter, con.identifier, pushRequest.Push.LedgerVersion, ignoreEvents)
 	}
 
 	proxiesConvergeDelay.Record(time.Since(pushRequest.Start).Seconds())
@@ -179,11 +178,11 @@ func (s *DiscoveryServer) receiveDelta(con *Connection, identities []string) {
 		req, err := con.deltaStream.Recv()
 		if err != nil {
 			if istiogrpc.IsExpectedGRPCError(err) {
-				deltaLog.Infof("ADS: %q %s terminated %v", con.PeerAddr, con.ConID, err)
+				deltaLog.Infof("ADS: %q %s terminated", con.peerAddr, con.identifier)
 				return
 			}
 			con.errorChan <- err
-			deltaLog.Errorf("ADS: %q %s terminated with error: %v", con.PeerAddr, con.ConID, err)
+			deltaLog.Errorf("ADS: %q %s terminated with error: %v", con.peerAddr, con.identifier, err)
 			totalXDSInternalErrors.Increment()
 			return
 		}
@@ -199,13 +198,13 @@ func (s *DiscoveryServer) receiveDelta(con *Connection, identities []string) {
 				return
 			}
 			defer s.closeConnection(con)
-			deltaLog.Infof("ADS: new delta connection for node:%s", con.ConID)
+			deltaLog.Infof("ADS: new delta connection for node:%s", con.identifier)
 		}
 
 		select {
 		case con.deltaReqChan <- req:
 		case <-con.deltaStream.Context().Done():
-			deltaLog.Infof("ADS: %q %s terminated with stream closed", con.PeerAddr, con.ConID)
+			deltaLog.Infof("ADS: %q %s terminated with stream closed", con.peerAddr, con.identifier)
 			return
 		}
 	}
@@ -237,7 +236,7 @@ func (conn *Connection) sendDelta(res *discovery.DeltaDiscoveryResponse) error {
 			conn.proxy.Unlock()
 		}
 	} else {
-		deltaLog.Infof("Timeout writing %s", conn.ConID)
+		deltaLog.Infof("Timeout writing %s", conn.identifier)
 		xdsResponseWriteTimeouts.Increment()
 	}
 	return err
@@ -247,7 +246,8 @@ func (conn *Connection) sendDelta(res *discovery.DeltaDiscoveryResponse) error {
 // handles 'push' requests and close - the code will eventually call the 'push' code, and it needs more mutex
 // protection. Original code avoided the mutexes by doing both 'push' and 'process requests' in same thread.
 func (s *DiscoveryServer) processDeltaRequest(req *discovery.DeltaDiscoveryRequest, con *Connection) error {
-	if !s.shouldProcessRequest(con.proxy, deltaToSotwRequest(req)) {
+	if req.TypeUrl == v3.HealthInfoType {
+		s.handleHealthRequest(con.proxy, deltaToSotwRequest(req))
 		return nil
 	}
 	if strings.HasPrefix(req.TypeUrl, v3.DebugType) {
@@ -256,7 +256,7 @@ func (s *DiscoveryServer) processDeltaRequest(req *discovery.DeltaDiscoveryReque
 			&model.PushRequest{Full: true, Push: con.proxy.LastPushContext})
 	}
 	if s.StatusReporter != nil {
-		s.StatusReporter.RegisterEvent(con.ConID, req.TypeUrl, req.ResponseNonce)
+		s.StatusReporter.RegisterEvent(con.identifier, req.TypeUrl, req.ResponseNonce)
 	}
 	shouldRespond := s.shouldRespondDelta(con, req)
 	if !shouldRespond {
@@ -292,7 +292,7 @@ func (s *DiscoveryServer) shouldRespondDelta(con *Connection, request *discovery
 	// will be different from the version sent. But it is fragile to rely on that.
 	if request.ErrorDetail != nil {
 		errCode := codes.Code(request.ErrorDetail.Code)
-		deltaLog.Warnf("ADS:%s: ACK ERROR %s %s:%s", stype, con.ConID, errCode.String(), request.ErrorDetail.GetMessage())
+		deltaLog.Warnf("ADS:%s: ACK ERROR %s %s:%s", stype, con.identifier, errCode.String(), request.ErrorDetail.GetMessage())
 		incrementXDSRejects(request.TypeUrl, con.proxy.ID, errCode.String())
 		if s.StatusGen != nil {
 			s.StatusGen.OnNack(con.proxy, deltaToSotwRequest(request))
@@ -317,7 +317,7 @@ func (s *DiscoveryServer) shouldRespondDelta(con *Connection, request *discovery
 	// We should always respond with the current resource names.
 	if previousInfo == nil {
 		// TODO: can we distinguish init and reconnect? Do we care?
-		deltaLog.Debugf("ADS:%s: INIT/RECONNECT %s %s", stype, con.ConID, request.ResponseNonce)
+		deltaLog.Debugf("ADS:%s: INIT/RECONNECT %s %s", stype, con.identifier, request.ResponseNonce)
 		con.proxy.Lock()
 		con.proxy.WatchedResources[request.TypeUrl] = &model.WatchedResource{
 			TypeUrl:       request.TypeUrl,
@@ -332,7 +332,7 @@ func (s *DiscoveryServer) shouldRespondDelta(con *Connection, request *discovery
 	// TODO: due to concurrent unsubscribe, this probably doesn't make sense. Do we need any logic here?
 	if request.ResponseNonce != "" && request.ResponseNonce != previousInfo.NonceSent {
 		deltaLog.Debugf("ADS:%s: REQ %s Expired nonce received %s, sent %s", stype,
-			con.ConID, request.ResponseNonce, previousInfo.NonceSent)
+			con.identifier, request.ResponseNonce, previousInfo.NonceSent)
 		xdsExpiredNonce.With(typeTag.Value(v3.GetMetricType(request.TypeUrl))).Increment()
 		con.proxy.Lock()
 		con.proxy.WatchedResources[request.TypeUrl].NonceNacked = ""
@@ -365,126 +365,13 @@ func (s *DiscoveryServer) shouldRespondDelta(con *Connection, request *discovery
 	// Envoy can send two DiscoveryRequests with same version and nonce
 	// when it detects a new resource. We should respond if they change.
 	if oldAck {
-		deltaLog.Debugf("ADS:%s: ACK  %s %s", stype, con.ConID, request.ResponseNonce)
+		deltaLog.Debugf("ADS:%s: ACK  %s %s", stype, con.identifier, request.ResponseNonce)
 		return false
 	}
 	deltaLog.Debugf("ADS:%s: RESOURCE CHANGE previous resources: %v, new resources: %v %s %s", stype,
-		previousResources, deltaResources, con.ConID, request.ResponseNonce)
+		previousResources, deltaResources, con.identifier, request.ResponseNonce)
 
 	return true
-}
-
-// Push an XDS resource for the given connection. Configuration will be generated
-// based on the passed in generator. Based on the updates field, generators may
-// choose to send partial or even no response if there are no changes.
-func (s *DiscoveryServer) pushDeltaXds(con *Connection,
-	w *model.WatchedResource, subscribe []string, req *model.PushRequest) error {
-	if w == nil {
-		return nil
-	}
-	gen := s.findGenerator(w.TypeUrl, con)
-	if gen == nil {
-		return nil
-	}
-	t0 := time.Now()
-
-	originalW := w
-	// If subscribe is set, client is requesting specific resources. We should just generate the
-	// new resources it needs, rather than the entire set of known resources.
-	if subscribe != nil {
-		w = &model.WatchedResource{
-			TypeUrl:       w.TypeUrl,
-			ResourceNames: subscribe,
-		}
-	}
-
-	var res model.Resources
-	var deletedRes model.DeletedResources
-	var logdata model.XdsLogDetails
-	var usedDelta bool
-	var err error
-	switch g := gen.(type) {
-	case model.XdsDeltaResourceGenerator:
-		res, deletedRes, logdata, usedDelta, err = g.GenerateDeltas(con.proxy, req, w)
-		if features.EnableUnsafeDeltaTest {
-			fullRes, _, _ := g.Generate(con.proxy, originalW, req)
-			s.compareDiff(con, originalW, fullRes, res, deletedRes, usedDelta, subscribe)
-		}
-	case model.XdsResourceGenerator:
-		res, logdata, err = g.Generate(con.proxy, w, req)
-	}
-	if err != nil || (res == nil && deletedRes == nil) {
-		// If we have nothing to send, report that we got an ACK for this version.
-		if s.StatusReporter != nil {
-			s.StatusReporter.RegisterEvent(con.ConID, w.TypeUrl, req.Push.LedgerVersion)
-		}
-		return err
-	}
-	defer func() { recordPushTime(w.TypeUrl, time.Since(t0)) }()
-	resp := &discovery.DeltaDiscoveryResponse{
-		ControlPlane: ControlPlane(),
-		TypeUrl:      w.TypeUrl,
-		// TODO: send different version for incremental eds
-		SystemVersionInfo: req.Push.PushVersion,
-		Nonce:             nonce(req.Push.LedgerVersion),
-		Resources:         res,
-	}
-	currentResources := extractNames(res)
-	if usedDelta {
-		resp.RemovedResources = deletedRes
-	} else if req.Full {
-		// similar to sotw
-		subscribed := sets.NewSet(w.ResourceNames...)
-		subscribed.Delete(currentResources...)
-		resp.RemovedResources = subscribed.SortedList()
-	}
-	if len(resp.RemovedResources) > 0 {
-		deltaLog.Debugf("ADS:%v %s REMOVE %v", v3.GetShortType(w.TypeUrl), con.ConID, resp.RemovedResources)
-	}
-	// normally wildcard xds `subscribe` is always nil, just in case there are some extended type not handled correctly.
-	if subscribe == nil && isWildcardTypeURL(w.TypeUrl) {
-		// this is probably a bad idea...
-		con.proxy.Lock()
-		w.ResourceNames = currentResources
-		con.proxy.Unlock()
-	}
-
-	configSize := ResourceSize(res)
-	configSizeBytes.With(typeTag.Value(w.TypeUrl)).Record(float64(configSize))
-
-	ptype := "PUSH"
-	info := ""
-	if logdata.Incremental {
-		ptype = "PUSH INC"
-	}
-	if len(logdata.AdditionalInfo) > 0 {
-		info = " " + logdata.AdditionalInfo
-	}
-	if err := con.sendDelta(resp); err != nil {
-		if recordSendError(w.TypeUrl, err) {
-			deltaLog.Warnf("%s: Send failure for node:%s resources:%d size:%s%s: %v",
-				v3.GetShortType(w.TypeUrl), con.proxy.ID, len(res), util.ByteCount(configSize), info, err)
-		}
-		return err
-	}
-
-	switch {
-	case logdata.Incremental:
-		if deltaLog.DebugEnabled() {
-			deltaLog.Debugf("%s: %s%s for node:%s resources:%d size:%s%s",
-				v3.GetShortType(w.TypeUrl), ptype, req.PushReason(), con.proxy.ID, len(res), util.ByteCount(configSize), info)
-		}
-	default:
-		debug := ""
-		if deltaLog.DebugEnabled() {
-			// Add additional information to logs when debug mode enabled.
-			debug = " nonce:" + resp.Nonce + " version:" + resp.SystemVersionInfo
-		}
-		deltaLog.Infof("%s: %s%s for node:%s resources:%d size:%v%s%s", v3.GetShortType(w.TypeUrl), ptype, req.PushReason(), con.proxy.ID, len(res),
-			util.ByteCount(ResourceSize(res)), info, debug)
-	}
-
-	return nil
 }
 
 func newDeltaConnection(peerAddr string, stream DeltaDiscoveryStream) *Connection {
@@ -492,8 +379,8 @@ func newDeltaConnection(peerAddr string, stream DeltaDiscoveryStream) *Connectio
 		pushChannel:  make(chan *Event),
 		initialized:  make(chan struct{}),
 		stop:         make(chan struct{}),
-		PeerAddr:     peerAddr,
-		Connect:      time.Now(),
+		peerAddr:     peerAddr,
+		connectedAt:  time.Now(),
 		deltaStream:  stream,
 		deltaReqChan: make(chan *discovery.DeltaDiscoveryRequest, 1),
 		errorChan:    make(chan error, 1),
